requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
idna==3.7
Mako==1.3.2
MarkupSafe==2.1.5
mkdocs-git-committers-plugin-2==2.2.3
mkdocs-git-revision-date-localized-plugin==1.2.4
mkdocs-material==9.5.6
mkdocs-material-extensions==1.3.1
mypy==1.9.0
mypy-extensions==1.0.0
platformdirs==4.2.0
pydantic==2.6.4
pydantic_core==2.16.3
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlalchemy-stubs==0.4
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0

pillow~=10.3.0
ultralytics~=8.2.14
aiofiles~=23.2.1
numpy~=1.26.4